Correção da tela de vitória, remoção de trecho desnecessário
Tela de vitória podia ser despausada com SPACEBAR, isso foi corrigido.
Painel de win separado do script de Pausa
Função antiga usada para delay do reset powerup foi removida por ter sido substituída por uma mais simples
