+++ SeaPong/Assets/Play/Scripts/Pause.cs
@@ -6,9 +6,7 @@ public class Pause : MonoBehaviour
 {
     public GameObject pausePanel;
     public GameObject pauseText;
-    public GameObject victoryPanel;
-    public GameObject MenuButton;
-    public TMP_Text victoryText;
+    public PainelController panelManager;
     private bool isPaused = false;
     public bool IsPaused
     {
@@ -25,19 +23,9 @@ public class Pause : MonoBehaviour
         Time.timeScale = IsPaused ? 0 : 1;
     }
 
-    public void EndGame(string message)
-    {
-        isPaused = true; // Pausa lógica do jogo
-        Time.timeScale = 0; // Garante que o jogo pare
-        victoryPanel.SetActive(isPaused);
-        MenuButton.SetActive(isPaused);
-        if (victoryText != null)
-            victoryText.text = message;
-    }
-
     void Update()
     {
-        if (Input.GetKeyDown(KeyCode.Space)) // Verifica se a tecla ESPAÇO foi pressionada
+        if (Input.GetKeyDown(KeyCode.Space) && !panelManager.GameWon) // Verifica se a tecla ESPAÇO foi pressionada
         {
             TogglePause();
         }
