+++ SeaPong/Assets/Settings/Build Profiles/New Linux Profile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5ec9bec92ef01ab41af0ac4a473348ba
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
